Phase 5 & 6: add scheduler, main entry point, and Streamlit dashboard
- src/scheduler.py: APScheduler BlockingScheduler, 6h interval, misfire_grace_time=600
- src/main.py: init DB, immediate pipeline run, SIGTERM handler with wait=True
- dashboard/app.py: metrics row, Human Queue (review + manual apply sections),
  Applied funnel, All Jobs with search/filter, Analytics with score histogram
  and applications-over-time chart; graceful DB error handling per section
- config/settings.py: fetch_interval_hours int → float for test flexibility
- CLAUDE.md: document PYTHONPATH=. requirement for local runs
- docs: mark phases 5 & 6 complete, document Human Queue two-section design

+++ config/settings.py
@@ -41,7 +41,7 @@ class Settings(BaseSettings):
     max_auto_applies_per_day: int = 20
     max_human_review_per_run: int = 5
 
-    fetch_interval_hours: int = 6
+    fetch_interval_hours: float = 6
     job_max_age_hours: int = 48
     zero_result_alert_threshold: int = 3
